# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * VeriGate identity verification API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SiteType = typeof SiteType[keyof typeof SiteType];


export const SiteType = {
  school: 'school',
  estate: 'estate',
} as const;

export interface Site {
  id: string;
  name: string;
  type: SiteType;
}

export type ProfileRole = typeof ProfileRole[keyof typeof ProfileRole];


export const ProfileRole = {
  guardian: 'guardian',
  backup_pickup: 'backup_pickup',
  resident: 'resident',
  visitor: 'visitor',
} as const;

export type ProfileStatus = typeof ProfileStatus[keyof typeof ProfileStatus];


export const ProfileStatus = {
  active: 'active',
  revoked: 'revoked',
} as const;

export interface Profile {
  id: string;
  siteId: string;
  name: string;
  role: ProfileRole;
  linkedTo: string;
  status: ProfileStatus;
  photoUrl: string;
  /** @nullable */
  faceDescriptor?: number[] | null;
  /** @nullable */
  visitorWindowStart?: string | null;
  /** @nullable */
  visitorWindowEnd?: string | null;
  createdAt: string;
}

export type ProfileInputRole = typeof ProfileInputRole[keyof typeof ProfileInputRole];


export const ProfileInputRole = {
  guardian: 'guardian',
  backup_pickup: 'backup_pickup',
  resident: 'resident',
  visitor: 'visitor',
} as const;

export interface ProfileInput {
  siteId: string;
  /** @minLength 2 */
  name: string;
  role: ProfileInputRole;
  /** @minLength 2 */
  linkedTo: string;
  photoUrl: string;
  /** @nullable */
  visitorWindowStart?: string | null;
  /** @nullable */
  visitorWindowEnd?: string | null;
  /** @nullable */
  faceDescriptor?: number[] | null;
}

export type ProfileStatusInputStatus = typeof ProfileStatusInputStatus[keyof typeof ProfileStatusInputStatus];


export const ProfileStatusInputStatus = {
  active: 'active',
  revoked: 'revoked',
} as const;

export interface ProfileStatusInput {
  status: ProfileStatusInputStatus;
}

export type VerificationDecisionInputResult = typeof VerificationDecisionInputResult[keyof typeof VerificationDecisionInputResult];


export const VerificationDecisionInputResult = {
  verified: 'verified',
  denied: 'denied',
  override: 'override',
} as const;

export interface VerificationDecisionInput {
  siteId: string;
  /** @nullable */
  matchedProfileId?: string | null;
  gateName: string;
  result: VerificationDecisionInputResult;
  /**
     * @minimum 0
     * @maximum 1
     */
  similarityScore: number;
  /** @minimum 0 */
  decisionMs: number;
  /** @nullable */
  faceDescriptor?: number[] | null;
  /** @nullable */
  overrideReason?: string | null;
  flaggedSuspicious?: boolean;
}

export type VerificationEventResult = typeof VerificationEventResult[keyof typeof VerificationEventResult];


export const VerificationEventResult = {
  verified: 'verified',
  denied: 'denied',
  override: 'override',
} as const;

export interface VerificationEvent {
  id: string;
  siteId: string;
  /** @nullable */
  matchedProfileId: string | null;
  gateName: string;
  result: VerificationEventResult;
  similarityScore: number;
  decisionMs: number;
  staffId: string;
  /** @nullable */
  overrideReason: string | null;
  flaggedSuspicious: boolean;
  createdAt: string;
}

export interface DashboardSummary {
  totalAttempts: number;
  verifiedHandoverRate: number;
  medianVerificationSpeed: number;
  overrideRate: number;
  flaggedCount: number;
  recentEvents: VerificationEvent[];
}

export type ListVerificationEventsParams = {
/**
 * @nullable
 */
siteId?: string | null;
/**
 * @nullable
 */
result?: ListVerificationEventsResult;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListVerificationEventsResult = typeof ListVerificationEventsResult[keyof typeof ListVerificationEventsResult] | null;


export const ListVerificationEventsResult = {
  verified: 'verified',
  denied: 'denied',
  override: 'override',
} as const;

export type GetDashboardSummaryParams = {
/**
 * @nullable
 */
siteId?: string | null;
};