# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * VeriGate identity verification API
 * OpenAPI spec version: 0.1.0
 */
import type { VerificationEvent } from './verificationEvent';

export interface DashboardSummary {
  totalAttempts: number;
  verifiedHandoverRate: number;
  medianVerificationSpeed: number;
  overrideRate: number;
  flaggedCount: number;
  recentEvents: VerificationEvent[];
}