# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * VeriGate identity verification API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileRole } from './profileRole';
import type { ProfileStatus } from './profileStatus';

export interface Profile {
  id: string;
  siteId: string;
  name: string;
  role: ProfileRole;
  linkedTo: string;
  status: ProfileStatus;
  photoUrl: string;
  /** @nullable */
  faceDescriptor?: number[] | null;
  /** @nullable */
  visitorWindowStart?: Date | null;
  /** @nullable */
  visitorWindowEnd?: Date | null;
  createdAt: Date;
}