# lib/api-zod/src/generated/types/profileStatusInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * VeriGate identity verification API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileStatusInputStatus } from './profileStatusInputStatus';

export interface ProfileStatusInput {
  status: ProfileStatusInputStatus;
}