# lib/api-zod/src/generated/types/site.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * VeriGate identity verification API
 * OpenAPI spec version: 0.1.0
 */
import type { SiteType } from './siteType';

export interface Site {
  id: string;
  name: string;
  type: SiteType;
}