# lib/api-zod/src/generated/types/verificationDecisionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * VeriGate identity verification API
 * OpenAPI spec version: 0.1.0
 */
import type { VerificationDecisionInputResult } from './verificationDecisionInputResult';

export interface VerificationDecisionInput {
  siteId: string;
  /** @nullable */
  matchedProfileId?: string | null;
  gateName: string;
  result: VerificationDecisionInputResult;
  /**
     * @minimum 0
     * @maximum 1
     */
  similarityScore: number;
  /** @minimum 0 */
  decisionMs: number;
  /** @nullable */
  faceDescriptor?: number[] | null;
  /** @nullable */
  overrideReason?: string | null;
  flaggedSuspicious?: boolean;
}